requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-mistralai>=0.0.1
langgraph>=0.0.10
duckduckgo-search>=4.1.1
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0 